Move class to relevant namespace
