+++ Unit4.Automation.Tests/CostCentreHierarchyTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Collections.Generic;
 using Unit4.Automation.Interfaces;
 using Unit4.Automation.Model;
+using Unit4.Automation.Commands.BcrCommand;
 
 namespace Unit4.Automation.Tests
 {
