+++ Unit4/Commands/BcrCommand/CostCentreHierarchy.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 using System.Linq;
 using Unit4.Automation.Interfaces;
 using Unit4.Automation.Model;
-using Unit4.Automation.Commands.BcrCommand;
+using Unit4.Automation;
 
-namespace Unit4.Automation
+namespace Unit4.Automation.Commands.BcrCommand
 {
     internal class CostCentreHierarchy
     {
